Refactors Selector / Match into ResigQuery
This change produces a ResigQuery type which is closer to the actual
jQuery type and splits it into partials for future expansion.  I expect
ResigQuery to be quite lengthy based on what I've seen from CsQuery.

A _lot_ of this work is inspired by CsQuery.  I have considered just
using CsQuery as the jQuery replacement, but like the underlying DOM
produced by AngleSharp.

https://github.com/jamietre/CsQuery
https://github.com/AngleSharp